+++ BSQ/includes/header.h
@@ -1,0 +1,12 @@
+#ifndef HEADER_H
+# define HEADER_H
+# define SIZE 4096
+# include <stdio.h>
+# include <fcntl.h>
+# include <stdlib.h>
+# include <unistd.h>
+
+int	map_check(char *str);
+int	map_lines(char *str);
+
+#endif
